+++ src/buffer.h
@@ -1,6 +1,6 @@
 /*
  *  mhuxd - mircoHam device mutliplexer/demultiplexer
- *  Copyright (C) 2012-2013  Matthias Moeller, DJ5QV
+ *  Copyright (C) 2012-2026  Matthias Moeller, DJ5QV
  *
  *  This program can be distributed under the terms of the GNU GPLv2.
  *  See the file COPYING
@@ -29,7 +29,7 @@ uint16_t buf_size_avail(struct buffer *b);
 void buf_add_size(struct buffer *b, uint16_t size_inc);
 void buf_consume(struct buffer *b, uint16_t size_inc);
 int buf_append_c(struct buffer *b, unsigned char c);
-int buf_append(struct buffer *b, const unsigned char *p, ssize_t len);
+int buf_append(struct buffer *b, const unsigned char *p, uint16_t len);
 int buf_append_str(struct buffer *b, const unsigned char *p);
 int buf_get_c(struct buffer *b);
 void buf_remove_front(struct buffer *b, uint16_t len);
